+++ android/src/main/java/com/drumpicker/DrumPickerViewManager.kt
@@ -42,6 +42,7 @@ class DrumPickerViewManager :
         "containerBackgroundColor" -> view.setContainerBackgroundColorProp(value)
         "itemBackgroundColor" -> view.setItemBackgroundColorProp(value)
         "hapticFeedback" -> view.setHapticFeedbackProp(value)
+        "circular" -> view.setCircularProp(value)
         "enableScrollByTapOnItem" -> view.setEnableScrollByTapOnItemProp(value)
         "onValueChangingEnabled" -> view.setOnValueChangingEnabledProp(value)
         else -> delegate.setProperty(view, key, value)
@@ -112,6 +113,10 @@ class DrumPickerViewManager :
 
   override fun setHapticFeedback(view: DrumPickerView?, value: Boolean) {
     view?.setHapticFeedbackProp(value)
+  }
+
+  override fun setCircular(view: DrumPickerView?, value: Boolean) {
+    view?.setCircularProp(value)
   }
 
   override fun setEnableScrollByTapOnItem(view: DrumPickerView?, value: Boolean) {